simplify retractions on powermanifold, continues documentation. Transers productmanifold.
